+++ src/app/locales/en/components/content-users.js
@@ -20,6 +20,7 @@ export default {
   profile: 'Profile',
   fullName: 'Full name',
   username: 'Username',
+  userId: 'User ID',
   usernameNoBasicAuthHint: 'A human-readable identifier, unique across the system, e.g. "johndoe13". Makes it easier for other users to find you.',
   usernameWithBasicAuthHint: 'A human-readable identifier, unique across the system, e.g. "johndoe13". Makes it easier for other users to find you and can be used along with password for authenticating.',
   password: 'Password',