+++ src/app/locales/en/components/sidebar-providers.js
@@ -2,5 +2,9 @@ export default {
   providerItem: {
     spaceCountHint: 'Number of supported spaces: {{count}}',
     totalSupportHint: 'Total support size: {{size}}',
+    providerId: 'provider ID',
+    copyIdAction: 'Copy provider ID',
+    providerDomain: 'provider domain',
+    copyDomainAction: 'Copy provider domain',
   },
 };